requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
pytz==2021.1
sqlparse==0.4.1
gunicorn
python-dotenv~=0.19.1
channels~=3.0.4
PyJWT~=2.1.0